refactor: split agent types, trim comments, use JSON.stringify for pixel params
